Update Panda ACC state tracking; other refactoring
Add CRC support for TSK_06, and have Panda watch that instead of ACC_06
for ACC state. ACC_06 can appear on bus 0 or 2 depending on how Panda is
wired to the car (at the driver assist camera or CAN gateway); the Panda
message timer/CRC/counter code doesn't make that easy to change at time
of fingerprint. TSK_06 gives us the same info and it's always on bus 0.

Other minor refactoring and comment updates.